refactor(tools): extract source test modules
- Move tool unit tests into sibling *_test.rs files
- Keep tool implementation files focused on runtime behavior
