+++ crates/aion-tools/src/glob.rs
@@ -143,113 +143,5 @@ impl Tool for GlobTool {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use serde_json::json;
-    use std::fs;
-    use std::path::PathBuf;
-    use tempfile::tempdir;
-
-    use aion_types::tool::ToolResult;
-
-    async fn run_glob(pattern: &str, path: &str) -> ToolResult {
-        let tool = GlobTool::new(PathBuf::from(path));
-        let input = json!({ "pattern": pattern, "path": path });
-        tool.execute(input).await
-    }
-
-    #[tokio::test]
-    async fn test_glob_matches_pattern() {
-        let dir = tempdir().unwrap();
-        let base = dir.path();
-
-        fs::write(base.join("main.rs"), "fn main() {}").unwrap();
-        fs::write(base.join("lib.rs"), "pub mod lib;").unwrap();
-        fs::write(base.join("notes.txt"), "some notes").unwrap();
-        fs::write(base.join("readme.md"), "# Readme").unwrap();
-
-        let result = run_glob("*.rs", base.to_str().unwrap()).await;
-
-        assert!(!result.is_error, "glob should succeed");
-        let lines: Vec<&str> = result.content.lines().collect();
-        assert_eq!(lines.len(), 2, "should match exactly 2 .rs files");
-        for line in &lines {
-            assert!(line.ends_with(".rs"), "each match should be a .rs file, got: {}", line);
-        }
-        assert!(!result.content.contains("notes.txt"), "should not include .txt files");
-        assert!(!result.content.contains("readme.md"), "should not include .md files");
-    }
-
-    #[tokio::test]
-    async fn test_glob_no_matches() {
-        let dir = tempdir().unwrap();
-        let base = dir.path();
-
-        fs::write(base.join("main.rs"), "fn main() {}").unwrap();
-        fs::write(base.join("lib.rs"), "pub mod lib;").unwrap();
-
-        let result = run_glob("*.xyz", base.to_str().unwrap()).await;
-
-        assert!(!result.is_error, "no-match glob should not be an error");
-        assert_eq!(result.content, "No files matched the pattern");
-    }
-
-    #[tokio::test]
-    async fn test_glob_with_limit() {
-        let dir = tempdir().unwrap();
-        let base = dir.path();
-
-        for i in 0..5 {
-            fs::write(base.join(format!("file_{}.txt", i)), format!("content {}", i)).unwrap();
-        }
-
-        let result = run_glob("*.txt", base.to_str().unwrap()).await;
-
-        assert!(!result.is_error, "glob should succeed");
-        let lines: Vec<&str> = result.content.lines().collect();
-        assert_eq!(lines.len(), 5, "all 5 files should be returned");
-    }
-
-    #[tokio::test]
-    async fn test_glob_recursive() {
-        let dir = tempdir().unwrap();
-        let base = dir.path();
-
-        // Create nested directory structure
-        let sub_a = base.join("a");
-        let sub_b = base.join("a").join("b");
-        fs::create_dir_all(&sub_b).unwrap();
-
-        fs::write(base.join("root.txt"), "root level").unwrap();
-        fs::write(sub_a.join("mid.txt"), "middle level").unwrap();
-        fs::write(sub_b.join("deep.txt"), "deep level").unwrap();
-        // Non-matching file
-        fs::write(sub_a.join("skip.rs"), "not a txt").unwrap();
-
-        let result = run_glob("**/*.txt", base.to_str().unwrap()).await;
-
-        assert!(!result.is_error, "recursive glob should succeed");
-        let lines: Vec<&str> = result.content.lines().collect();
-        assert_eq!(lines.len(), 3, "should find 3 .txt files across all levels");
-        assert!(result.content.contains("root.txt"), "should include root-level file");
-        assert!(result.content.contains("mid.txt"), "should include mid-level file");
-        assert!(result.content.contains("deep.txt"), "should include deep-level file");
-        assert!(!result.content.contains("skip.rs"), "should not include .rs files");
-    }
-
-    #[tokio::test]
-    async fn execute_uses_cwd_for_relative_path() {
-        let tmp = tempdir().unwrap();
-        fs::write(tmp.path().join("marker.txt"), "hello").unwrap();
-
-        let tool = GlobTool::new(tmp.path().to_path_buf());
-        let input = json!({"pattern": "marker.txt"});
-        let result = tool.execute(input).await;
-        assert!(!result.is_error, "unexpected error: {}", result.content);
-        assert!(
-            result.content.contains("marker.txt"),
-            "should find marker.txt, got: {}",
-            result.content
-        );
-    }
-}
+#[path = "glob_test.rs"]
+mod glob_test;
