+++ crates/aion-tools/src/lib.rs
@@ -82,43 +82,5 @@ pub trait Tool: Send + Sync {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn truncate_utf8_ascii_within_limit() {
-        assert_eq!(truncate_utf8("hello", 80), "hello");
-    }
-
-    #[test]
-    fn truncate_utf8_ascii_at_boundary() {
-        assert_eq!(truncate_utf8("abcde", 3), "abc");
-    }
-
-    #[test]
-    fn truncate_utf8_multibyte_snaps_back() {
-        // '些' is 3 bytes (E4 BA 9B) starting at index 79 would span 79..82
-        let s = "# 用 script 模拟 TTY 交互来添加 DeepSeek 提供商\n# 首先看看有哪些";
-        let result = truncate_utf8(s, 80);
-        assert!(result.len() <= 80);
-        assert!(result.is_char_boundary(result.len()));
-    }
-
-    #[test]
-    fn truncate_utf8_empty() {
-        assert_eq!(truncate_utf8("", 80), "");
-    }
-
-    #[test]
-    fn truncate_utf8_zero_limit() {
-        assert_eq!(truncate_utf8("hello", 0), "");
-    }
-
-    #[test]
-    fn truncate_utf8_emoji() {
-        // 🦀 is 4 bytes
-        let s = "aaa🦀bbb";
-        assert_eq!(truncate_utf8(s, 4), "aaa");
-        assert_eq!(truncate_utf8(s, 7), "aaa🦀");
-    }
-}
+#[path = "lib_test.rs"]
+mod lib_test;
