+++ crates/aion-tools/src/read_test.rs
@@ -1,0 +1,244 @@
+use super::*;
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use serde_json::json;
+    use std::io::Write;
+    use tempfile::tempdir;
+
+    use aion_config::file_cache::FileCacheConfig;
+
+    fn make_cache() -> Arc<RwLock<FileStateCache>> {
+        let config = FileCacheConfig {
+            max_entries: 100,
+            max_size_bytes: 25 * 1024 * 1024,
+            enabled: true,
+        };
+        Arc::new(RwLock::new(FileStateCache::new(&config)))
+    }
+
+    // -- Basic read tests (no cache) --
+
+    #[tokio::test]
+    async fn test_read_file_full() {
+        let dir = tempdir().unwrap();
+        let file_path = dir.path().join("test.txt");
+        let mut file = std::fs::File::create(&file_path).unwrap();
+        writeln!(file, "line one").unwrap();
+        writeln!(file, "line two").unwrap();
+        writeln!(file, "line three").unwrap();
+        drop(file);
+
+        let tool = ReadTool::new(None);
+        let input = json!({ "file_path": file_path.to_str().unwrap() });
+        let result = tool.execute(input).await;
+
+        assert!(!result.is_error);
+        assert!(result.content.contains("1\tline one"));
+        assert!(result.content.contains("2\tline two"));
+        assert!(result.content.contains("3\tline three"));
+    }
+
+    #[tokio::test]
+    async fn test_read_file_with_offset_and_limit() {
+        let dir = tempdir().unwrap();
+        let file_path = dir.path().join("lines.txt");
+        let mut file = std::fs::File::create(&file_path).unwrap();
+        for i in 1..=10 {
+            writeln!(file, "line {}", i).unwrap();
+        }
+        drop(file);
+
+        let tool = ReadTool::new(None);
+        let input = json!({
+            "file_path": file_path.to_str().unwrap(),
+            "offset": 2,
+            "limit": 3
+        });
+        let result = tool.execute(input).await;
+
+        assert!(!result.is_error);
+        let lines: Vec<&str> = result.content.lines().collect();
+        assert_eq!(lines.len(), 3);
+        assert!(lines[0].contains("3\tline 3"));
+        assert!(lines[1].contains("4\tline 4"));
+        assert!(lines[2].contains("5\tline 5"));
+    }
+
+    #[tokio::test]
+    async fn test_read_nonexistent_file() {
+        let tool = ReadTool::new(None);
+        let input = json!({ "file_path": "/tmp/nonexistent_file_abc123.txt" });
+        let result = tool.execute(input).await;
+
+        assert!(result.is_error);
+        assert!(result.content.contains("Failed to read file"));
+    }
+
+    #[tokio::test]
+    async fn test_read_empty_file() {
+        let dir = tempdir().unwrap();
+        let file_path = dir.path().join("empty.txt");
+        std::fs::File::create(&file_path).unwrap();
+
+        let tool = ReadTool::new(None);
+        let input = json!({ "file_path": file_path.to_str().unwrap() });
+        let result = tool.execute(input).await;
+
+        assert!(!result.is_error);
+        assert!(result.content.is_empty());
+    }
+
+    #[tokio::test]
+    async fn test_read_large_file_truncation() {
+        let dir = tempdir().unwrap();
+        let file_path = dir.path().join("large.txt");
+        let mut file = std::fs::File::create(&file_path).unwrap();
+        for i in 1..=200 {
+            writeln!(file, "line number {}", i).unwrap();
+        }
+        drop(file);
+
+        let tool = ReadTool::new(None);
+        let input = json!({ "file_path": file_path.to_str().unwrap() });
+        let result = tool.execute(input).await;
+
+        assert!(!result.is_error);
+        let lines: Vec<&str> = result.content.lines().collect();
+        assert_eq!(lines.len(), 200);
+        assert!(lines[0].contains("1\tline number 1"));
+        assert!(lines[199].contains("200\tline number 200"));
+    }
+
+    // -- Dedup tests (with cache) --
+
+    #[tokio::test]
+    async fn dedup_returns_stub_on_unchanged_file() {
+        let dir = tempdir().unwrap();
+        let file_path = dir.path().join("dedup.txt");
+        std::fs::write(&file_path, "hello\n").unwrap();
+
+        let cache = make_cache();
+        let tool = ReadTool::new(Some(cache));
+
+        let input = json!({ "file_path": file_path.to_str().unwrap() });
+
+        // First read: full content.
+        let r1 = tool.execute(input.clone()).await;
+        assert!(!r1.is_error);
+        assert!(r1.content.contains("hello"));
+
+        // Second read: dedup stub.
+        let r2 = tool.execute(input).await;
+        assert!(!r2.is_error);
+        assert_eq!(r2.content, FILE_UNCHANGED_STUB);
+    }
+
+    #[tokio::test]
+    async fn dedup_returns_new_content_after_modification() {
+        let dir = tempdir().unwrap();
+        let file_path = dir.path().join("modified.txt");
+        std::fs::write(&file_path, "version1\n").unwrap();
+
+        let cache = make_cache();
+        let tool = ReadTool::new(Some(cache));
+
+        let input = json!({ "file_path": file_path.to_str().unwrap() });
+
+        let r1 = tool.execute(input.clone()).await;
+        assert!(r1.content.contains("version1"));
+
+        // Modify the file — ensure mtime changes.
+        std::thread::sleep(std::time::Duration::from_millis(50));
+        std::fs::write(&file_path, "version2\n").unwrap();
+
+        let r2 = tool.execute(input).await;
+        assert!(!r2.is_error);
+        assert!(r2.content.contains("version2"));
+    }
+
+    #[tokio::test]
+    async fn dedup_different_offset_limit_returns_full() {
+        let dir = tempdir().unwrap();
+        let file_path = dir.path().join("multi.txt");
+        let mut file = std::fs::File::create(&file_path).unwrap();
+        for i in 1..=20 {
+            writeln!(file, "line {}", i).unwrap();
+        }
+        drop(file);
+
+        let cache = make_cache();
+        let tool = ReadTool::new(Some(cache));
+
+        let input1 = json!({
+            "file_path": file_path.to_str().unwrap(),
+            "offset": 0,
+            "limit": 10
+        });
+        let r1 = tool.execute(input1).await;
+        assert!(!r1.is_error);
+        assert!(r1.content.contains("line 1"));
+
+        // Different range: should return full content, not stub.
+        let input2 = json!({
+            "file_path": file_path.to_str().unwrap(),
+            "offset": 10,
+            "limit": 10
+        });
+        let r2 = tool.execute(input2).await;
+        assert!(!r2.is_error);
+        assert!(r2.content.contains("line 11"));
+        assert!(!r2.content.contains(FILE_UNCHANGED_STUB));
+    }
+
+    #[tokio::test]
+    async fn no_cache_always_returns_full_content() {
+        let dir = tempdir().unwrap();
+        let file_path = dir.path().join("nocache.txt");
+        std::fs::write(&file_path, "data\n").unwrap();
+
+        let tool = ReadTool::new(None);
+        let input = json!({ "file_path": file_path.to_str().unwrap() });
+
+        let r1 = tool.execute(input.clone()).await;
+        assert!(r1.content.contains("data"));
+
+        let r2 = tool.execute(input).await;
+        assert!(r2.content.contains("data"));
+        assert_ne!(r2.content, FILE_UNCHANGED_STUB);
+    }
+
+    #[tokio::test]
+    async fn nonexistent_file_not_cached() {
+        let cache = make_cache();
+        let tool = ReadTool::new(Some(cache.clone()));
+
+        let input = json!({ "file_path": "/tmp/nonexistent_xyz_789.txt" });
+        let r = tool.execute(input).await;
+        assert!(r.is_error);
+
+        // Cache should be empty.
+        let c = cache.read().unwrap();
+        assert!(c.is_empty());
+    }
+
+    #[tokio::test]
+    async fn dedup_empty_file() {
+        let dir = tempdir().unwrap();
+        let file_path = dir.path().join("empty.txt");
+        std::fs::File::create(&file_path).unwrap();
+
+        let cache = make_cache();
+        let tool = ReadTool::new(Some(cache));
+
+        let input = json!({ "file_path": file_path.to_str().unwrap() });
+
+        let r1 = tool.execute(input.clone()).await;
+        assert!(!r1.is_error);
+
+        let r2 = tool.execute(input).await;
+        assert!(!r2.is_error);
+        assert_eq!(r2.content, FILE_UNCHANGED_STUB);
+    }
+}
