+++ crates/aion-tools/src/registry.rs
@@ -64,254 +64,5 @@ impl ToolRegistry {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::Tool;
-    use aion_protocol::events::ToolCategory;
-    use aion_types::tool::ToolResult;
-    use async_trait::async_trait;
-
-    /// A minimal Tool implementation used only in tests
-    struct MockTool {
-        tool_name: String,
-        tool_description: String,
-        tool_category: ToolCategory,
-    }
-
-    #[async_trait]
-    impl Tool for MockTool {
-        fn name(&self) -> &str {
-            &self.tool_name
-        }
-
-        fn description(&self) -> &str {
-            &self.tool_description
-        }
-
-        fn input_schema(&self) -> serde_json::Value {
-            serde_json::json!({"type": "object"})
-        }
-
-        fn is_concurrency_safe(&self, _input: &serde_json::Value) -> bool {
-            true
-        }
-
-        async fn execute(&self, _input: serde_json::Value) -> ToolResult {
-            ToolResult {
-                content: "ok".to_string(),
-                is_error: false,
-            }
-        }
-
-        fn category(&self) -> ToolCategory {
-            self.tool_category
-        }
-    }
-
-    /// Helper to create a MockTool with the given name and description
-    fn make_tool(name: &str, description: &str) -> Box<MockTool> {
-        Box::new(MockTool {
-            tool_name: name.to_string(),
-            tool_description: description.to_string(),
-            tool_category: ToolCategory::Info,
-        })
-    }
-
-    fn make_tool_with_category(name: &str, description: &str, category: ToolCategory) -> Box<MockTool> {
-        Box::new(MockTool {
-            tool_name: name.to_string(),
-            tool_description: description.to_string(),
-            tool_category: category,
-        })
-    }
-
-    #[test]
-    fn test_register_and_get() {
-        let mut registry = ToolRegistry::new();
-        registry.register(make_tool("my_tool", "does something"));
-
-        let found = registry.get("my_tool");
-        assert!(found.is_some(), "registered tool should be retrievable by name");
-        assert_eq!(found.unwrap().name(), "my_tool");
-    }
-
-    #[test]
-    fn test_get_nonexistent_returns_none() {
-        let registry = ToolRegistry::new();
-
-        let result = registry.get("ghost");
-        assert!(result.is_none(), "looking up an unregistered name should return None");
-    }
-
-    #[test]
-    fn test_tool_names() {
-        let mut registry = ToolRegistry::new();
-        registry.register(make_tool("alpha", "first tool"));
-        registry.register(make_tool("beta", "second tool"));
-        registry.register(make_tool("gamma", "third tool"));
-
-        let mut names = registry.tool_names();
-        names.sort(); // sort for a stable assertion order
-        assert_eq!(names, vec!["alpha", "beta", "gamma"]);
-    }
-
-    #[test]
-    fn test_to_tool_defs() {
-        let mut registry = ToolRegistry::new();
-        registry.register(make_tool("tool_a", "description A"));
-        registry.register(make_tool("tool_b", "description B"));
-
-        let defs = registry.to_tool_defs();
-        assert_eq!(
-            defs.len(),
-            2,
-            "to_tool_defs should return one entry per registered tool"
-        );
-
-        // Collect (name, description) pairs for assertion independent of order
-        let mut pairs: Vec<(&str, &str)> = defs.iter().map(|d| (d.name.as_str(), d.description.as_str())).collect();
-        pairs.sort();
-
-        assert_eq!(pairs[0], ("tool_a", "description A"));
-        assert_eq!(pairs[1], ("tool_b", "description B"));
-
-        // Verify the input_schema field is populated correctly
-        let expected_schema = serde_json::json!({"type": "object"});
-        for def in &defs {
-            assert_eq!(def.input_schema, expected_schema);
-        }
-    }
-
-    // --- to_tool_defs_filtered tests ---
-
-    #[test]
-    fn filtered_by_category_returns_matching_tools() {
-        let mut registry = ToolRegistry::new();
-        registry.register(make_tool_with_category("Read", "read files", ToolCategory::Info));
-        registry.register(make_tool_with_category("Write", "write files", ToolCategory::Edit));
-        registry.register(make_tool_with_category(
-            "ExecCommand",
-            "run commands",
-            ToolCategory::Exec,
-        ));
-        registry.register(make_tool_with_category(
-            "ExitPlanMode",
-            "exit plan mode",
-            ToolCategory::Info,
-        ));
-
-        let defs = registry.to_tool_defs_filtered(|t| t.category() == ToolCategory::Info);
-
-        let names: Vec<&str> = defs.iter().map(|d| d.name.as_str()).collect();
-        assert!(names.contains(&"Read"));
-        assert!(names.contains(&"ExitPlanMode"));
-        assert!(!names.contains(&"Write"));
-        assert!(!names.contains(&"ExecCommand"));
-    }
-
-    #[test]
-    fn filtered_by_name_excludes_specific_tool() {
-        let mut registry = ToolRegistry::new();
-        registry.register(make_tool("alpha", "first"));
-        registry.register(make_tool("beta", "second"));
-        registry.register(make_tool("gamma", "third"));
-
-        let defs = registry.to_tool_defs_filtered(|t| t.name() != "beta");
-
-        let names: Vec<&str> = defs.iter().map(|d| d.name.as_str()).collect();
-        assert_eq!(names.len(), 2);
-        assert!(names.contains(&"alpha"));
-        assert!(names.contains(&"gamma"));
-        assert!(!names.contains(&"beta"));
-    }
-
-    #[test]
-    fn filtered_accept_all_matches_to_tool_defs() {
-        let mut registry = ToolRegistry::new();
-        registry.register(make_tool("a", "tool a"));
-        registry.register(make_tool("b", "tool b"));
-
-        let all = registry.to_tool_defs();
-        let filtered = registry.to_tool_defs_filtered(|_| true);
-
-        assert_eq!(all.len(), filtered.len());
-        for (a, f) in all.iter().zip(filtered.iter()) {
-            assert_eq!(a.name, f.name);
-        }
-    }
-
-    #[test]
-    fn filtered_reject_all_returns_empty() {
-        let mut registry = ToolRegistry::new();
-        registry.register(make_tool("a", "tool a"));
-
-        let defs = registry.to_tool_defs_filtered(|_| false);
-        assert!(defs.is_empty());
-    }
-
-    #[test]
-    fn filtered_empty_registry_returns_empty() {
-        let registry = ToolRegistry::new();
-        let defs = registry.to_tool_defs_filtered(|_| true);
-        assert!(defs.is_empty());
-    }
-
-    // --- deferred flag tests ---
-
-    /// A minimal Tool that overrides is_deferred() to return true
-    struct DeferredMockTool {
-        tool_name: String,
-    }
-
-    #[async_trait]
-    impl Tool for DeferredMockTool {
-        fn name(&self) -> &str {
-            &self.tool_name
-        }
-
-        fn description(&self) -> &str {
-            "a deferred tool"
-        }
-
-        fn input_schema(&self) -> serde_json::Value {
-            serde_json::json!({"type": "object", "properties": {"x": {"type": "string"}}})
-        }
-
-        fn is_concurrency_safe(&self, _input: &serde_json::Value) -> bool {
-            true
-        }
-
-        fn is_deferred(&self) -> bool {
-            true
-        }
-
-        async fn execute(&self, _input: serde_json::Value) -> ToolResult {
-            ToolResult {
-                content: "ok".to_string(),
-                is_error: false,
-            }
-        }
-
-        fn category(&self) -> ToolCategory {
-            ToolCategory::Info
-        }
-    }
-
-    #[test]
-    fn to_tool_defs_includes_deferred_flag() {
-        let mut registry = ToolRegistry::new();
-        registry.register(make_tool("core_tool", "a core tool"));
-        let defs = registry.to_tool_defs();
-        assert!(!defs[0].deferred, "default tools should not be deferred");
-    }
-
-    #[test]
-    fn to_tool_defs_deferred_tool_flagged() {
-        let mut registry = ToolRegistry::new();
-        registry.register(Box::new(DeferredMockTool {
-            tool_name: "lazy_tool".to_string(),
-        }));
-        let defs = registry.to_tool_defs();
-        assert!(defs[0].deferred, "deferred tool should have deferred=true");
-    }
-}
+#[path = "registry_test.rs"]
+mod registry_test;
