+++ crates/aion-tools/src/tool_search_test.rs
@@ -1,0 +1,78 @@
+use super::*;
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    fn build_tool_defs() -> Vec<ToolDef> {
+        vec![
+            ToolDef {
+                name: "Read".into(),
+                description: "Read a file".into(),
+                input_schema: json!({"type": "object", "properties": {"path": {"type": "string"}}}),
+                deferred: false,
+            },
+            ToolDef {
+                name: "SpawnTool".into(),
+                description: "Spawn sub-agents".into(),
+                input_schema: json!({"type": "object", "properties": {"agents": {"type": "array"}}}),
+                deferred: true,
+            },
+            ToolDef {
+                name: "EnterPlanMode".into(),
+                description: "Enter plan mode".into(),
+                input_schema: json!({"type": "object", "properties": {}}),
+                deferred: true,
+            },
+        ]
+    }
+
+    #[tokio::test]
+    async fn search_by_exact_name() {
+        let tool = ToolSearchTool::new(build_tool_defs());
+        let result = tool.execute(json!({"query": "SpawnTool"})).await;
+        assert!(!result.is_error);
+        assert!(result.content.contains("SpawnTool"));
+        assert!(result.content.contains("Spawn sub-agents"));
+        assert!(result.content.contains("parameters"));
+    }
+
+    #[tokio::test]
+    async fn search_case_insensitive() {
+        let tool = ToolSearchTool::new(build_tool_defs());
+        let result = tool.execute(json!({"query": "spawntool"})).await;
+        assert!(!result.is_error);
+        assert!(result.content.contains("SpawnTool"));
+    }
+
+    #[tokio::test]
+    async fn search_by_description_keyword() {
+        let tool = ToolSearchTool::new(build_tool_defs());
+        let result = tool.execute(json!({"query": "plan"})).await;
+        assert!(!result.is_error);
+        assert!(result.content.contains("EnterPlanMode"));
+    }
+
+    #[tokio::test]
+    async fn search_excludes_non_deferred() {
+        let tool = ToolSearchTool::new(build_tool_defs());
+        let result = tool.execute(json!({"query": "Read"})).await;
+        // "Read" is not deferred, should not appear in results
+        assert!(!result.content.contains("\"name\": \"Read\"") || result.content.contains("No deferred tools"));
+    }
+
+    #[tokio::test]
+    async fn search_no_match() {
+        let tool = ToolSearchTool::new(build_tool_defs());
+        let result = tool.execute(json!({"query": "nonexistent"})).await;
+        assert!(!result.is_error);
+        assert!(result.content.contains("No deferred tools"));
+    }
+
+    #[tokio::test]
+    async fn search_empty_query_returns_error() {
+        let tool = ToolSearchTool::new(build_tool_defs());
+        let result = tool.execute(json!({"query": ""})).await;
+        assert!(result.is_error);
+    }
+}
